Fixing HBase 2.* minicluster test There were a few problems - IntegrationTests did not include TestListTables - There was a race condition when multiple subclasses of AbstractTestListTables ran together relating to listing tables via a pattern. - Multiple async admin methods throw `UnsupportedOperationException` in HBase's Admin implementations.  Those tests should only run in Bigtable mode. - HBase minicluster gets into a really bad state when this is run `asyncAdmin.getDescriptor(null).get()`.  The minicluster never shuts down.  This test should only run in Bigtable mode.

+++ bigtable-hbase-2.x-parent/bigtable-hbase-2.x-integration-tests/src/test/java/com/google/cloud/bigtable/hbase/TestListTablesHBase2.java
@@ -61,7 +61,7 @@ public class TestListTablesHBase2 extends AbstractTestListTables {
   
   @Override
   protected void deleteTable(Admin admin, TableName tableName) throws Exception {
-    if (enableAsyncDelete) {
+    if (enableAsyncDelete && sharedTestEnv.isBigtable()) {
       admin.disableTableAsync(tableName).get();
       admin.deleteTableAsync(tableName).get();
     } else {
